Moved plot docker to studio module
