+++ studio/demo/plot_docker_demo.cpp
@@ -28,7 +28,7 @@ TESSERACT_COMMON_IGNORE_WARNINGS_PUSH
 #include <QDebug>
 TESSERACT_COMMON_IGNORE_WARNINGS_POP
 
-#include <tesseract_qt/plot/plot_docker.h>
+#include <tesseract_qt/studio/plot_docker.h>
 
 int main(int argc, char** argv)
 {
