+++ studio/include/tesseract_qt/studio/plot_docker.h
