+++ studio/include/tesseract_qt/studio/plot_docker_toolbar.ui
